+++ js/viz/landscape_ist.js
@@ -220,7 +220,9 @@ export const landscape_ist = async (
   max_tiles_to_view = 50,
   scale_bar_microns_per_pixel = null,
   base_urls = [],
-  cell_name_prefix = false
+  cell_name_prefix = false,
+  centroids = {},
+  use_adata_3d_centroids = false
 ) => {
   if (width === 0) {
     width = '100%';
@@ -399,6 +401,12 @@ export const landscape_ist = async (
     viz_state.umap.has_umap = true;
   }
   viz_state.umap.umap = umap;
+
+  viz_state.centroids3d = {
+    has_centroids: Object.keys(centroids).length > 0,
+    centroids,
+    requested: use_adata_3d_centroids,
+  };
 
   const isUmapInit = landscape_state === 'umap';
   viz_state.obs_store.umap_state.set(isUmapInit);